+++ logouthandler.go
@@ -27,11 +27,11 @@ func (h *LogoutRedirectionHandler) ServeHTTP(w http.ResponseWriter, r *http.Requ
 		http.Redirect(w, r, h.logoutUrl, http.StatusSeeOther)
 		return
 	}
+
 	h.delegate.ServeHTTP(w, r)
 }
 
 func (h *LogoutRedirectionHandler) isLogoutRequest(r *http.Request) bool {
-	log.Infof("Inspecting request %s url %s", r.Method, r.URL)
 	return (h.logoutMethod != "" || h.logoutPath != "") &&
 		(h.logoutMethod == "" || r.Method == h.logoutMethod) &&
 		(h.logoutPath == "" || h.logoutPath != "" && strings.HasSuffix(r.URL.Path, h.logoutPath))