+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- [#14] add ability to bypass CAS-authentication for certain non-browser-requests
+  - This prevents request-throttling in CAS for requests that only have dogu-internal authentication  
+- [#14] add throttling for bypassed-requests
 
 ## [v1.1.0] - 2020-09-04
 ### Added